web/requirements_web.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
plotly>=5.17.0
pandas>=2.2.0
numpy>=1.26.0
aiofiles>=23.2.0
requests>=2.31.0
PyJWT>=2.10.1
